--- CryptoRandom.cs
@@ -1,9 +1,0 @@
-﻿namespace kov.NET.Protections
-{
-    internal class CryptoRandom
-    {
-        public CryptoRandom()
-        {
-        }
-    }
-}
