+++ Protections/Analyzers/DefAnalyzer.cs
