+++ Protections/Analyzers/EventDefAnalyzer.cs
