+++ Protections/Analyzers/FieldDefAnalyzer.cs
