+++ Protections/Analyzers/MethodDefAnalyzer.cs
