+++ Protections/Analyzers/PropertyDefAnalyzer.cs
