+++ Protections/Utils/CryptoRandom.cs
@@ -1,0 +1,9 @@
+﻿namespace kov.NET.Protections
+{
+    internal class CryptoRandom
+    {
+        public CryptoRandom()
+        {
+        }
+    }
+}
