+++ Protections/Utils/InjectHelper.cs
