+++ packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="dnlib" version="3.3.2" targetFramework="net472" />
+  <package id="dnlib" version="3.3.5" targetFramework="net48" />
 </packages>
